savePlace edits

+++ backend/api.js
@@ -41,7 +41,7 @@ exports.setApp = function ( app, client )
 
         try
         {
-            const db = await client.db();
+            const db = client.db();
 
             // grabs folder based on folderId.
             const result = await db.collection('Folders').findOne
@@ -77,7 +77,8 @@ exports.setApp = function ( app, client )
     {
         var token = require('./createJWT.js'); var msg = ''; var error = '';
 
-        const jwToken = req.body.jwToken; const uid = req.body.userId;
+        const jwToken = req.body.jwToken; 
+        const uid = req.body.userId;
         const fid = req.body.folderId;
 
         // Checks if the JWT is expired
@@ -99,24 +100,24 @@ exports.setApp = function ( app, client )
 
         try
         {
-            const db = await client.db();
+            const db = client.db();
             // array of folders matching the userId.
             const result = await db.collection('Folders').aggregate([
                 {
                   $match: {
-                    "userId": 3,
-                    "folderId": 53
+                    "userId": uid,
+                    "folderId": fid
                   }
                 },
                 {
-                  "$project": {
-                    "placeList": 1,
+                  $project: {
+                    placeList: 1,
                     _id: 0
                   }
                 }
               ]).toArray();
-
-            msg = result;
+              
+            msg = result[0].placeList;
         }
         catch(e)
         {
@@ -169,7 +170,7 @@ exports.setApp = function ( app, client )
         console.log({jwToken, newFolderName, fid})
         try
         {
-            const db = await client.db();
+            const db = client.db();
 
             // edits name of folder that matches userId and folderId.
             const result = await db.collection('Folders').updateOne
@@ -213,7 +214,6 @@ exports.setApp = function ( app, client )
 
         // no need for folder schema since we are just inserting into an existing folder. 
         const fid = req.body.folderId;
-        const uid = req.body.userId;
 
         // using place schema.
         const newPlace = new Place
@@ -224,6 +224,7 @@ exports.setApp = function ( app, client )
 
         var msg = '';
         var error = '';
+
 
         // Checks if the JWT is expired
         // Sets the error and returns
@@ -242,19 +243,6 @@ exports.setApp = function ( app, client )
             return;
         }
         const db = await client.db();
-        const folderExist = await db.collection('Folders').find({userId: uid,folderId:fid}).toArray();
-      
-
-        if (folderExist.length === 0)
-        {
-            // Send an error that the userId doesn't exist
-            error = "no folder with given userId and folderId";
-            ret = { error: error };
-            res.status(500).json(ret);
-
-            // Exit the api call
-            return;
-        }
         try
         {
 
@@ -262,7 +250,7 @@ exports.setApp = function ( app, client )
             // if didn't add because of duplicate: message.modifiedCount = 0 and message.matchedCount = 1.
             const result = await db.collection('Folders').updateOne
             (
-                {userId: uid, folderId: fid},
+                {folderId: fid},
                 {$addToSet: 
                     {placeList: 
                         {
@@ -326,7 +314,7 @@ exports.setApp = function ( app, client )
             console.log(e.message);
             return;
         }
-        const db = await client.db();
+        const db = client.db();
         try
         {
             
@@ -397,13 +385,26 @@ exports.setApp = function ( app, client )
             console.log(e.message);
             return;
         }
+        const db = client.db();
+        const folderExist = await db.collection('Folders').find({folderId:thisFolder}).toArray();
+      
+
+        if (folderExist.length === 0)
+        {
+            // Send an error that the userId doesn't exist
+            error = "folder doesn't exist";
+            ret = { error: error };
+            res.status(500).json(ret);
+
+            // Exit the api call
+            return;
+        }
         // Actual folder deletion
         try
         {
-            const db = await client.db()
+           
             const result = await db.collection('Folders').deleteOne({folderId: thisFolder});
             msg = result;
-            
         }
         catch(e)
         {
@@ -456,7 +457,7 @@ exports.setApp = function ( app, client )
 
         try
         {
-            const db = await client.db();
+            const db = client.db();
             const result = await db.collection('Users').deleteOne({userId: thisUserId, password: thisUserPass});
             msg = result;
         }
@@ -503,7 +504,7 @@ exports.setApp = function ( app, client )
             return;
         }
 
-        const db = await client.db();
+        const db = client.db();
 
         // Search the database for the username and email
         const usernameExist = await db.collection('Users').find({username:newUsername}).toArray();
@@ -885,7 +886,7 @@ exports.setApp = function ( app, client )
         var ret;
 
         // Connect to the database
-        const db = await client.db();
+        const db = client.db();
 
         // Search the database for the username and email
         const userExist = await db.collection('Users').find({username:newUser.username}).toArray();
@@ -970,7 +971,7 @@ exports.setApp = function ( app, client )
 
         const { username, password } = req.body;
 
-        const db = await client.db();
+        const db = client.db();
         const results = await db.collection('Users').findOne({username:username,password:password});
 
         var id = -1;
@@ -1042,7 +1043,7 @@ exports.setApp = function ( app, client )
         {
             console.log(e.message);
         }
-        const db = await client.db();
+        const db = client.db();
         const userIdExist = await db.collection('Users').find({userId:newFolder.userId}).toArray();
       
 
@@ -1154,7 +1155,7 @@ exports.setApp = function ( app, client )
         const userId = req.body.userId;
 
         // Connect to the database
-        const db = await client.db();
+        const db = client.db();
         // Search the database for the userId
         const userEmail = await db.collection('Users').findOne({userId:userId});
 
@@ -1247,7 +1248,7 @@ exports.setApp = function ( app, client )
         console.log(req.query);
 
         // Connect to the database
-        const db = await client.db();
+        const db = client.db();
         const checkedUser = await db.collection('Users').findOne({email: req.query.email});
         const confirmUser = await db.collection('Users').updateOne({userId: checkedUser.userId}, {$set: {emailConfirm: 1}, $unset: {confirmToken: ''}});
         console.log(confirmUser);
@@ -1311,7 +1312,7 @@ exports.setApp = function ( app, client )
         try
         {
             // Connect to the database
-            const db = await client.db();
+            const db = client.db();
             // Search the database for the userId
             const userConfirm = await db.collection('Users').findOne({userId:userId});
 
@@ -1366,7 +1367,7 @@ exports.setApp = function ( app, client )
         // SEND EMAIL STUFF
 
         // Connect to the database
-        const db = await client.db();
+        const db = client.db();
         // Search the database for the account tied to the email
         const resetUser = await db.collection('Users').findOne({email:req.body.email});
         console.log(resetUser);
@@ -1477,14 +1478,12 @@ exports.setApp = function ( app, client )
         console.log(req.query);
 
         // Connect to the database
-        const db = await client.db();
+        const db = client.db();
         const checkedUser = await db.collection('Users').findOne({email: req.query.email});
-        const confirmUser = await db.collection('Users').findOne({userId: checkedUser.userId});
-        console.log("reset user:");
+        const confirmUser = await db.collection('Users').updateOne({userId: checkedUser.userId}, {$unset: {confirmToken: ''}});
         console.log(confirmUser);
 
-        // Send token and redirect to PasswordChange
-        
+ 
         // NEEDS TO REDIRECT TO REACT SERVER
         res.redirect(`${req.query.redirect}/PasswordChange`);
         //////
@@ -1498,9 +1497,6 @@ exports.setApp = function ( app, client )
 
         // Error message to be sent
         var error = '';
-
-        // Response message to be sent
-        var responseMsg = '';
 
         // Variable to store the response to be sent back to front-end
         const r = 
@@ -1547,7 +1543,7 @@ exports.setApp = function ( app, client )
         const userId = req.body.userId;
 
         // Connect to the database
-        const db = await client.db();
+        const db = client.db();
         // Search the database for the userId and update password
         const resetUser = await db.collection('Users').findOne({userId:userId});
         const confirmUser = await db.collection('Users').updateOne({userId: resetUser.userId}, 
@@ -1592,7 +1588,7 @@ exports.setApp = function ( app, client )
     var results;
     try
     {
-        const db = await client.db();
+        const db = client.db();
         results = await db.collection('Folders').find({userId:userId, folderType:folderType}).toArray();
 
     }
